chore: remove resume page
