--- src/pages/Resume.module.scss
@@ -1,219 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Open+Sans:wght@300;400&display=swap");
-
-$mobile-breakpoint: 500px;
-
-.resumeBody {
-  --color-teal-600: #006c75;
-  --color-gray-100: #c8c9ca;
-  --color-gray-200: #919394;
-  --color-gray-400: #656869;
-  --color-gray-700: #232728;
-  --color-dark-700: #a5bbbd;
-
-  --font-family-base: "Open Sans", sans-serif;
-  --font-family-headings: "Open Sans", sans-serif;
-
-  --main-background-color: white;
-
-  @media print {
-    background-color: white;
-  }
-
-  margin: 0;
-  padding: 0;
-  font-family: var(--font-family-base);
-  font-size: 12px;
-  color: var(--color-gray-700);
-  background-color: var(--main-background-color);
-  h1,
-  h2 {
-    font-family: var(--font-family-headings);
-    color: var(--color-gray-700);
-  }
-
-  h1 {
-    font-size: 3.75em;
-    font-weight: 300;
-  }
-
-  h1 span:first-of-type {
-    font-weight: lighter;
-  }
-
-  h2 {
-    font-size: 2em;
-    margin: 0 0 0.15em 0;
-    font-weight: 300;
-  }
-
-  h3 {
-    font-size: 1.15em;
-    font-weight: 300;
-    margin: 0.5em 0 0 0;
-    color: var(--color-gray-700);
-    &.bold {
-      font-weight: bold;
-    }
-  }
-
-  h4 {
-    margin: 0.25em 0 0 0;
-    color: var(--color-gray-700);
-    &.bold {
-      font-weight: bold;
-    }
-  }
-
-  a {
-    color: var(--color-teal-600);
-  }
-
-  p {
-    margin-bottom: 0.5em;
-    line-height: 1.5;
-  }
-
-  hr {
-    color: var(--color-gray-200);
-    &.mediumThickness {
-      border: 1px solid;
-    }
-  }
-
-  section {
-    margin: 1em 0;
-    padding: 0 0.25em;
-    &:first-of-type {
-      margin-bottom: 0;
-      margin-top: 0;
-      padding: 0;
-    }
-    &:nth-of-type(2) {
-      margin-top: 0;
-      padding: 0;
-    }
-    &:last-of-type {
-      margin: 1em 0 0 0;
-    }
-  }
-
-  .all-caps {
-    text-transform: uppercase;
-  }
-
-  .center-vertical {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-  }
-
-  .text-center {
-    text-align: center;
-  }
-
-  .bold {
-    font-weight: bold;
-  }
-
-  .no-bullets {
-    list-style-type: none;
-    padding-left: 0;
-    ul {
-      list-style-type: disc;
-    }
-  }
-
-  .flex {
-    display: flex;
-    &.justify-space-between {
-      justify-content: space-between;
-    }
-    &.direction-column {
-      flex-direction: column;
-    }
-  }
-
-  .header {
-    display: grid;
-    grid-template-columns: 55% 45%;
-    align-items: center;
-    padding: 1em;
-    background-color: var(--color-teal-300);
-    h1,
-    h2 {
-      color: var(--text-color-on-teal-300);
-      margin: 0;
-    }
-  }
-
-  .basicsInfo {
-    background-color: var(--color-teal-300);
-    padding: 0 1em 1em 1em;
-    ul {
-      justify-content: flex-start;
-      padding-left: 0;
-      li {
-        margin-left: 0;
-      }
-    }
-    a {
-      color: var(--text-color-on-teal-300);
-    }
-  }
-
-  .listHorizontal {
-    list-style: none;
-    display: flex;
-    justify-content: center;
-    flex-wrap: wrap;
-    li {
-      margin: 0 1em;
-    }
-  }
-
-  .listHorizontalBullets {
-    padding-left: 0;
-    li {
-      display: inline-block;
-      &:before {
-        content: "\00a0\2022\00a0\00a0";
-        color: #999;
-        color: rgba(0, 0, 0, 0.5);
-        font-size: 12px;
-      }
-      &:first-child:before {
-        content: "";
-      }
-    }
-  }
-
-  .skills-grid {
-    &:not(:first-child) {
-      border-top: 1px dashed var(--color-gray-200);
-    }
-    display: grid;
-    grid-template-columns: 110px 1fr;
-    align-items: center;
-    h3 {
-      text-transform: capitalize;
-    }
-  }
-  .skills-list {
-    list-style-type: none;
-    display: flex;
-    flex-wrap: wrap;
-    font-size: 0.8em;
-    padding: 0;
-    &.justify-center {
-      justify-content: center;
-    }
-  }
-
-  .skills-list li {
-    padding: 0.35em 0.75em;
-    border: solid 1px var(--color-teal-300);
-    border-radius: 10px;
-    margin: 0.25em;
-    height: 2.25em;
-  }
-}
